adding floorplan and ajax calls to get data from ws

+++ src/main/kotlin/org/techndrome/dogtracker/service/NoiseGenService.kt
@@ -26,7 +26,7 @@ class NoiseGenService {
     }
 
     fun moveDog(dog: Dog) {
-        dog.coord.x = rand.nextDouble()
-        dog.coord.y = rand.nextDouble()
+        dog.coord.x = rand.nextInt()
+        dog.coord.y = rand.nextInt()
     }
 }
